+++ packages/core/src/fst/index.ts
@@ -1,6 +1,6 @@
 // https://en.wikipedia.org/wiki/Finite-state_transducer
 
-import { State, StateMachine } from '.';
+import type { State, StateMachine } from '..';
 
 export type FSTTransition<TState, TInput, TOutput> = (
   state: TState,
@@ -19,18 +19,3 @@ export type FSTFrom<
 > = TMachine extends StateMachine<infer TContext, any, infer TEvent>
   ? FST<State<TContext, TEvent>, TEvent, any[]>
   : never;
-
-export function machineToFST<TMachine extends StateMachine<any, any, any>>(
-  machine: TMachine,
-  nextEvents?: FSTFrom<TMachine>['nextEvents']
-): FSTFrom<TMachine> {
-  return {
-    transition: ((state, input) => {
-      const nextState = machine.transition(state, input);
-      return [nextState, []];
-    }) as FSTFrom<TMachine>['transition'],
-    initialState: machine.initialState,
-    events: [] as FSTFrom<TMachine>['events'],
-    nextEvents
-  } as FSTFrom<TMachine>;
-}
